moved messages and types from utilities to root
