+++ src/Model/Logic/ILogic.ts
@@ -1,4 +1,4 @@
-import { Grid } from '../../Utilities/Types';
+import { Grid } from '../../Types';
 
 interface ILogic{
   toggleCellState(grid: Grid, x:number, y:number): Grid;
