+++ src/Model/Model.ts
@@ -1,5 +1,5 @@
 import State from './State/State';
-import Messages from '../Utilities/Messages';
+import Messages from '../Messages';
 import Observer from '../Utilities/Observer';
 import IModel from './IModel';
 import bindDecorator from 'bind-decorator';
